Avoid ambiguous overload error with GCC 3.3 on Mac OS X
git-svn-id: svn://svn.code.sf.net/p/fuego/code/trunk@619 2e953b5c-c64d-0410-be54-f773e93e544c

+++ smartgame/SgUctBiasTermPrecomp.cpp
@@ -23,7 +23,7 @@ SgUctBiasTermPrecomp::SgUctBiasTermPrecomp(std::size_t maxPosCount,
     for (size_t posCount = 0; posCount < maxPosCount; ++posCount)
         for (size_t moveCount = 0; moveCount < maxMoveCount; ++moveCount)
             Value(posCount, moveCount) =
-                Compute(log(posCount), moveCount);
+                Compute(log(static_cast<float>(posCount)), moveCount);
 }
 
 //----------------------------------------------------------------------------
